+++ packages/terra-abstract-modal/src/terra-dev-site/test/abstract-modal/AbstractModalIsFullscreen.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import AbstractModal from '../../../AbstractModal';
-import './AbstractModalTestStyles.scss';
+import './AbstractModalTestStyles.module.scss';
 
 class ModalIsFullscreen extends React.Component {
   constructor() {