+++ packages/terra-form-select/src/_Tag.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import styles from './_Tag.module.scss';
+import styles from './_Tag.module.module.scss';
 
 const cx = classNames.bind(styles);