+++ packages/terra-signature/src/Signature.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import 'terra-base/lib/baseStyles';
-import styles from './Signature.scss';
+import styles from './Signature.module.scss';
 
 const LINEWIDTHS = {
   EXTRA_FINE: 1,