+++ packages/terra-status/src/Status.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
-import styles from './Status.scss';
+import styles from './Status.module.scss';
 
 const cx = classNames.bind(styles);