+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -7,6 +7,8 @@ Unreleased
 * Resolve issue where disabled component doesn't disable time clarification
 
 ### Changed
+* Add .module extension to SCSS file(s)
+### Changed
 * Use datepicker that consumes new form inputs.
 
 2.22.0 - (July 17, 2018)